requirements.txt
# Research & Report Agent — Python dependencies
# Install: pip install -r requirements.txt

# ── Core LangChain stack ──────────────────────────────────────────────────────
langchain>=0.3.0,<1.0.0
langchain-groq>=0.2.0,<1.0.0
langchain-community>=0.3.0,<1.0.0
langchainhub>=0.1.21          # hub.pull() for ReAct prompts

# ── LLM provider ─────────────────────────────────────────────────────────────
groq>=0.11.0

# ── Vector memory ─────────────────────────────────────────────────────────────
chromadb>=0.5.0
sentence-transformers>=3.0.0  # for HuggingFaceEmbeddings (all-MiniLM-L6-v2)
tiktoken>=0.7.0

# ── Search tools ─────────────────────────────────────────────────────────────
duckduckgo-search>=6.0.0
ddgs>=9.0.0                   # required by langchain DuckDuckGoSearchRun backend
tavily-python>=0.3.0          # optional — better results than DuckDuckGo

# ── Web scraping ─────────────────────────────────────────────────────────────
requests>=2.32.0
beautifulsoup4>=4.12.0

# ── API server ────────────────────────────────────────────────────────────────
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic>=2.7.0

# ── Config & utilities ────────────────────────────────────────────────────────
python-dotenv>=1.0.0

# ── Testing ───────────────────────────────────────────────────────────────────
pytest>=8.0.0
pytest-asyncio>=0.23.0
